+++ core/urlAccess.py
@@ -18,7 +18,6 @@ from core.trace import TraceLevel, Trace
 import base64
 import config
 import json
-import requests
 import socket
 import ssl
 import sys
@@ -26,6 +25,9 @@ import time
 import traceback
 import urllib.request, urllib.error
 import warnings
+
+import requests
+from requests_toolbelt.multipart.encoder import MultipartEncoder
 
 ################################################################################
 # UrlStatus
@@ -69,7 +71,7 @@ class UrlAccess():
 
     @classmethod
     def process_push(self, redfishConfig, link, filename, payload = None):
-        Trace.log(TraceLevel.TRACE, '   ++ UrlAccess: process_push - ({}) session ({}:{})'.format(link.url, Label.decode(config.sessionIdVariable), redfishConfig.sessionKey))
+        Trace.log(TraceLevel.INFO, '++ UrlAccess: process_push - ({}) session ({}:{})'.format(link.url, Label.decode(config.sessionIdVariable), redfishConfig.sessionKey))
 
         warnings.filterwarnings('ignore', message='Unverified HTTPS request')
 
@@ -77,31 +79,68 @@ class UrlAccess():
 
         s = requests.Session()
 
-        files = [
-            ('downloadfile', (filename, open(filename, 'rb'), 'application/octet-stream'))
-        ]
+        Trace.log(TraceLevel.INFO, '++ UrlAccess: process_push - filename ({})'.format(filename))
+
+        files = {
+            ('UpdateFile', (filename, open(filename, 'rb'), 'application/octet-stream'))
+        }
 
         if redfishConfig.get_bool('httpbasicauth'):
             Trace.log(TraceLevel.DEBUG, '   -- Using HTTP Basic Auth')
             uername_password = redfishConfig.get_value('username') + ':' + redfishConfig.get_value('password')
             encoded = base64.b64encode(str.encode(uername_password))
             Trace.log(TraceLevel.DEBUG, '   -- uername_password is ({}) encoded is ({})'.format(uername_password, encoded))
-            s.headers.update({'Authorization': 'Basic ' + str(encoded)})
+            s.headers.update({'Authorization': 'Basic ' + str(encoded), })
         else:
             Trace.log(TraceLevel.INFO, '   -- Using X-Auth-Token: {}'.format(redfishConfig.sessionKey))
+            
+            # s.headers.update({'X-Auth-Token': redfishConfig.sessionKey, 'Content-Type': 'application/json; charset=utf-8'})
             s.headers.update({'X-Auth-Token': redfishConfig.sessionKey})
 
         try:
             fullUrl = redfishConfig.get_value('http') + '://' + redfishConfig.get_value('mcip') + link.url
+            Trace.log(TraceLevel.INFO, '   -- fullUrl: {}'.format(fullUrl))
+
             if payload is None:
+                Trace.log(TraceLevel.INFO, '   @@ post no JSON')
                 response = s.request('POST', fullUrl, files=files)
             else:
+                Trace.log(TraceLevel.INFO, '   @@ post w/ JSON')
                 if (redfishConfig.get_bool('dumppostdata')):
                     Trace.log(TraceLevel.INFO, '[[ POST DATA ({}) ]]'.format(link.url))
                     print(json.dumps(payload, indent=4))
                     Trace.log(TraceLevel.INFO, '[[ POST DATA END ]]')
-                # payload2 = json.dumps(payload, indent=4)
-                response = s.request('POST', fullUrl, data=payload, files=files, verify=False)
+                # response = s.request('POST', fullUrl, data=payload, files=files, verify=False)
+                multipart_req = requests.Request('POST', fullUrl, files=files, data=payload).prepare()
+
+            try:
+                Trace.log(TraceLevel.INFO, '===== A =====')
+                print(multipart_req.body.decode('utf-8'))
+                Trace.log(TraceLevel.INFO, '===== = =====')
+            except Exception as e:
+                Trace.log(TraceLevel.ERROR, 'EXCEPTION (1): {}'.format(e))
+
+
+            Trace.log(TraceLevel.INFO, '============================== NEW B ==============================')
+            m = MultipartEncoder(
+                fields={'Targets': '[]', '@Redfish.OperationApplyTime': 'Immediate',
+                        'UpdateFile': ('filename', open(filename, 'rb'), 'application/octet-stream')}
+            )
+
+            multipart_req = requests.Request('POST', fullUrl, data=m, headers={'Content-Type': m.content_type}).prepare()
+            print(multipart_req.body)
+            # r = requests.post('http://httpbin.org/post', data=m, headers={'Content-Type': m.content_type})
+            Trace.log(TraceLevel.INFO, '============================== END B ==============================')
+
+            response = s.request('POST', fullUrl, data=m, headers={'Content-Type': m.content_type}, verify=False)
+            # response = s.request('POST', fullUrl, data=payload, files=files, verify=False)
+
+            Trace.log(TraceLevel.INFO, '============================== NEW RESPONSE ==============================')
+            print(response.text)
+            print('status_code: {}'.format(response.status_code))
+            print('reason: {}'.format(response.reason))
+            print('request: {}'.format(response.request))
+            Trace.log(TraceLevel.INFO, '============================== END RESPONSE ==============================')
 
             link.elapsedMicroseconds = (time.time() - startTime) * 1000000
             link.urlData = response.text
@@ -227,13 +266,14 @@ class UrlAccess():
             # Print the contents of the HTTP message response
             read_op = getattr(err, "read", None)
             if (callable(read_op)):
-                Trace.log(TraceLevel.VERBOSE, '   ' + '='*60 + '  HTTP Error START  ' + '='*60)
+                Trace.log(TraceLevel.INFO, '')
+                Trace.log(TraceLevel.INFO, '='*120)
                 errorMessage = err.read()
                 if (redfishConfig.get_bool('dumphttpdata')):
-                    Trace.log(TraceLevel.INFO, '   -- httpData {}'.format(errorMessage))
+                    Trace.log(TraceLevel.INFO, 'httpData: {}'.format(errorMessage))
                 else:
-                    Trace.log(TraceLevel.VERBOSE, '  errorMessage = {}'.format(errorMessage))
-                Trace.log(TraceLevel.VERBOSE, '   ' + '='*60 + '  HTTP Error END  ' + '='*60)
+                    Trace.log(TraceLevel.INFO, 'errorMessage = {}'.format(errorMessage))
+                Trace.log(TraceLevel.INFO, '='*120)
                 pass
 
         except urllib.error.HTTPError as err: